[Perf/Interpreter]: Remove some yield cases from opcodes
Some opcodes had special handling for yield but this wasn't actually
possible to be hit (to my knowledge) for example ExecEq/Add/... all
checked whether or not the instructionPtr would be < 0 and if so threw
exceptions so having another check checking if the value is == -99 after
that doesn't make any sense...

This is unlikely to significantly improve performance because well it's
just another conditional on each loop but given the last one had about a
50ms impact I would expect something around that.

There are still some cases that check it (where it's possible to be hit)
but I'll probably remove those too.

I want to change how coroutines are handled anyways for performance
reasons (they should operate more like a c# coroutine).