[Perf]: Inline Stack Ops
I want to ideally look towards maybe even getting rid of the stack class
but for now forcing inline should give us most of the value.  I also
removed some of the IList stuff.

There is some slicing done but that is done inefficiently so I'll move
it to a proper slicer.